Update README.md
UIImageView's image property is UIImage? and you'd need to unwrap it before passing to the init method.
Many tend to trust Xcode and put ! after image, thinking it can never be nil (speaking from experience). 
But it can be, if image isn't loaded into the cell yet. 
Anyway, this change makes the code in the readme compilable and helps avoid stupid mistakes.